+++ src/constants/dbTypes.php
@@ -7,8 +7,5 @@ namespace BPT\constants;
  */
 class dbTypes {
     public const JSON = 'json';
-    /** @var string Will be added soon */
     public const MYSQL = 'mysql';
-    /** @var string Will be added soon */
-    public const MEDOO = 'medoo';
 }